Add abstractions for non-Gregorian calendars such as Ethiopian (#1505)
Calendars other than the Gregorian calendar can now be added more easily. These are introduced as appearances on the existing date types meaning that the underlying date is always Gregorian but the user sees pickers that follow the rules of the desired calendar.

The Ethiopian calendar is introduced through the ethiopian appearance. The implementation uses the Ethiopic joda Chronology. This appearance can be combined with others such as month-year or year.